Add MixedMetaWinner strategy.
A variation of MetaWinnerEnsemble strategy.  Basically, a team is created where on each turn a distribution is calculated based on the scores of the players.  This distribution is then used to randomly choose the next move. The strategy is then more likely to pick a winning team. This can also be seen as a variation on the MetaMixer strategy as it updates the mixed strategy on each turn.